pyproject.toml
[project]
name = "phpprof"
version = "0.1.0"
description = "Assistant pédagogique IA local pour PHP et Symfony"
authors = [
    { name = "Abdel TOUATI" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9",
    "huggingface-hub>=0.20",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=3.0",
    "requests>=2.31",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.8",
]
train = [
    "mlx>=0.16",
    "mlx-lm>=0.18",
]
eval = [
    "nltk>=3.8",
]
all = [
    "phpprof[dev,train,eval]",
]

[project.scripts]
phpprof = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
